Allow only specific examples to be compiled
In order to not change the existing behaviour of --with-examples,
this adds a new option that you can use instead:

  ./waf configure --with-example standard_mfcc,streaming_gfcc

Will build only the specified examples. You can specify one or more.
To compile all examples, use

  ./waf configure --with-examples

as before. If we are happy to change the behaviour of this option we
could require an argument such as 'all' to compile all examples.
optparse does not support an option that can have an optional argument,
so we cannot keep the existing behaviour.