New option to build dependency tree for a given algorithm
Update show_algo_dependencies script with a new option to build a
tree of dependencies given an algorithm. Currently the tree is not
output, only the recursive list of dependencies.

The recursive search is slow and can be better implemented by storign
dependencies previously analyzed before to a global variable.