unitSum normalization; better param names in HPCP
Add unitSum normalization (closes #348). Print warning when trying to
apply unitSum together with band preset (this combination was not tested).

Some code cleanup.

Rename 'splitFrequency' parameter to 'bandSplitFrequency'. This way it
is easier to see in documentation that it is related to bandPreset. Update
extractors using this parameter.

Better description for non-linear post processing in the doc.