Fix memory deallocation for python streaming mode
Fix a memory leak: the algorithms would not be deleted when the
corresponding python's algorithm object went out of scope in the
streaming mode.

Implement a new method deleteNetwork that deletes all algorithms
connected into a network given its root.

Add new option to visibleDependencies() method to omit warnings
when there are any unconnected outputs left in the network.