Only link against libessentia in the pc file. Fixes #340
When we want to link libessentia.so into a program we only need to
specify -lessentia. Other dependencies of essentia are already
correctly linked against the .so. If a user wants to also link
their program against the same dependency (e.g. libav) then they
should explicitly link against that themselves.