Faster VECTOR_VECTOR_REAL conversion to python types
Update the type wrapper for VECTOR_VECTOR_REAL to output a list of
1-dimensional numpy arrays instead of a list of lists.

Previously, a Python list of lists of floats was output in the case when
the inner vectors of an input VECTOR_VECTOR_REAL are of unequal size.
This resulted in a very slow behavior when those lists were further
passed to other algorithms.