Ensure no nan/inf values in kurtosis and skewness
Ensure that there won't be any nan/inf values due to division by
zero in helper functions for kurtosis and skewness.

Use hardcoded skewness=0 and kurtosis=-3 in the case of  zero spread
(similar to how it is done in the DistributionShape algorithm).

Remove explicit nan/inf check for kurtosis() and skewness() methods,
fix hardcoded value for kurtosis from 0 to -3.

Closes #689.