Fix unit test and document known Fails
testSpectrumSizeSmallerThanNumberOfBands was failing because of an
incorrect implementation.
The new way of computing the bands correctly assignates the central
frequencies: 57.26, 163.98, 469.57, 1344.60, 3850.23 and 11024.99 Hz
in that particular case.
This means that only the last band contains one spectral bin.
However, this was making the test fail, as it was designed to pass
only if no spectral bins were contained inside any band.
The test was modified to pass on the current implementation.