Put all Cephes code inside the namespace
In the previous commit only the methods in the header were into the
namespece.
This caused an error for multiple definitions of `gamma' in some
circumstances.
This commits puts all the library code into the namespace and adds its
gamma function to the header so it doesn't get confused with other
implementations.