Add related algorithms to each algorithm doc page
The documentation page for each algorithm now contains links to
related algorithms in a new section "See also". It is automatically
generated by a naive string matching for algorithm names in the
description.

This may result in some false positives (strings recognized as
algorithms while they are not), but that should not be a big deal
(e.g., see GeometricMean).

A link to alternative mode for the algorithm (standard-streaming)
is also included.