Update PitchYin testRealCase
The audio source was changed to a shorter monophonic signal.
The expected values are now stored in a .npy so they are lighter and
faster to load.
Although the output was compared to the vamp implementation and seen to
be very close the expected values were obtained from the algorithm
itself. This way the allowed error could be reduced so the assertion
will be triggered with any small variation from the current values.