Add TensorFlow headers an pkg-config files
generate_pc.py locates the python package for TensorFlow
and creates a pkg-config file to link agaist the .so libs
on it.
TODO: setup_tensorflow.sh should run previous script and
put the .pc file on an appropiate place.
Also make sure that Python is able to locate .so libs on
run time.