Allow using default paths together with custom PKG_CONFIG_PATH
PKG_CONFIG_LIBDIR was automatically assigned equal to PKG_CONFIG_PATH
if the latter was define. This made impossible to use a custom path
together with default pkg-config search paths.

Changed the behavior so that the --pkg-config-path sets both
PKG_CONFIG_PATH and PKG_CONFIG_LIBDIR, while a PKG_CONFIG_PATH can
be also passed on its own:
     PKG_CONFIG_PATH=... python3 waf configure

Improved configuration outputs. It is now explicitly reported with
the default paths are used or not.

The name of the --pkg-config-path flag is still somewhat confusing, so
we might want to rename it for something better in the future.