Restrict --whole-archive directive for tensorflow
When building Essentia statically to generate static executables, we
need the --whole-archive directive for libtensorflow-core so the linker
doesn't throw away the Ops objects (layers) not being used, as they will
be loaded dynamically on runtime via the protobuff files description.
By adding --no-whole-archive after libtensorflow-core we prevent the
rest of the dependencies to include all the symbols too.
This dramatically reduces the weight of the final executables, (from
300MB to 150MB with the default Tensorflow Ops).