Replace 'is not' by '!=' for literals
The compiler now produces a SyntaxWarning when identity checks (is and
is not) are used with certain types of literals (e.g. strings, numbers).
These can often work by accident in CPython, but are not guaranteed by
the language spec. The warning advises users to use equality tests
(== and !=) instead. (Contributed by Serhiy Storchaka in bpo-34850.)
https://bugs.python.org/issue34850