Centralize empty graphFilename handling
In `TensorFlowPredcit` the configuration is skipped when `graphFilename` is not specified.
This is because as we do not provide an standardized path for our models there is not suitable default value for it.

This commit extends this skip functionality to the case in which `graphFilename` is set to an empty string ("").
The main benefit is that algorithms depending on `TensorFlowPredcit` can default `graphFilename` to an empty string and leverage the configuration skip logic on `TensorflowPredict` instead of reimplementing it.
This also simplifies the logic required to enable loading frozen models (`graphFilename`) and TensorFlow 2 SavedModels (`savedModel`). (WIP)