Use a stronger threshold in tensorflowpredictmaest
This was achieves by making sure that the input melspectrogram was
computed exactly in the same way in TensorflowPredictMAEST and the
original pytorch implementation, using MonoLoader with
resampleQuality=1.

While using resampleQuality=4 has a numerical impact in the results, it
is not enough to change the prediction ranking according to informal
observations (could not be the same in corner cases).